refactor(explorer): migrate MATRIX_ELEMENT_SCAN to PanelBuilder registry
Phase 2/8 (4th of 9 PlotTypes total).

* New explorer_internals/matrix_element_sweep.py hosts
  MatrixElementSweepPanelBuilder + the absorbed
  display_matrixelement_sweep render function.
* Registry adds PlotType.MATRIX_ELEMENT_SCAN; the matching
  dispatch branches are removed from Explorer.build_panel and
  ExplorerSettings.build_settings_ui.

The settings UI here is byte-identical to MATRIX_ELEMENTS; a
follow-on commit during Phase 6 cleanup will dedupe the two
build_settings_ui bodies into a shared helper.
