+++ scqubits/explorer/explorer_internals/__init__.py
@@ -10,6 +10,9 @@ from scqubits.explorer.explorer_internals._base import PanelBuilder
 from scqubits.explorer.explorer_internals.energy_spectrum import (
     EnergySpectrumPanelBuilder,
 )
+from scqubits.explorer.explorer_internals.matrix_element_sweep import (
+    MatrixElementSweepPanelBuilder,
+)
 from scqubits.explorer.explorer_internals.matrix_elements import (
     MatrixElementsPanelBuilder,
 )
@@ -23,6 +26,7 @@ __all__ = [
     "PanelBuilder",
     "EnergySpectrumPanelBuilder",
     "MatrixElementsPanelBuilder",
+    "MatrixElementSweepPanelBuilder",
     "WavefunctionsPanelBuilder",
 ]
 
@@ -35,4 +39,5 @@ PANEL_BUILDERS: dict[PlotType, type[PanelBuilder]] = {
     PlotType.ENERGY_SPECTRUM: EnergySpectrumPanelBuilder,
     PlotType.WAVEFUNCTIONS: WavefunctionsPanelBuilder,
     PlotType.MATRIX_ELEMENTS: MatrixElementsPanelBuilder,
+    PlotType.MATRIX_ELEMENT_SCAN: MatrixElementSweepPanelBuilder,
 }
